+++ src/main/java/ch/hearc/zookeeper/entity/Equipment.java
@@ -17,7 +17,7 @@ public class Equipment
 	@Id
 	@GeneratedValue(strategy=GenerationType.AUTO)
 	@Column
-	private Integer id;
+	private long id;
 
 	@Column
 	private String name;
@@ -39,12 +39,12 @@ public class Equipment
 		this.sector = sector;
 	}
 
-	public Integer getId() 
+	public long getId() 
 	{
 		return id;
 	}
 
-	public void setId(Integer id) 
+	public void setId(long id) 
 	{
 		this.id = id;
 	}
