+++ src/main/java/ch/hearc/zookeeper/entity/Sector.java
@@ -14,17 +14,17 @@ public class Sector
 	@Id
 	@GeneratedValue(strategy=GenerationType.AUTO)
 	@Column
-	private Integer id;
+	private long id;
 
 	@Column
 	private String name;
 
-	public Integer getId() 
+	public long getId() 
 	{
 		return id;
 	}
 
-	public void setId(Integer id) 
+	public void setId(long id) 
 	{
 		this.id = id;
 	}
