+++ src/main/java/ch/hearc/zookeeper/entity/Stock.java
@@ -17,7 +17,7 @@ public class Stock
 	@Id
 	@GeneratedValue(strategy=GenerationType.AUTO)
 	@Column
-	private Integer id;
+	private long id;
 	
 	@Column
 	private int quantity;
@@ -26,12 +26,12 @@ public class Stock
     @JoinColumn(name = "id", nullable = false, insertable = false, updatable = false)
     private Equipment equipment;
 
-	public Integer getId() 
+	public long getId() 
 	{
 		return id;
 	}
 
-	public void setId(Integer id) 
+	public void setId(long id) 
 	{
 		this.id = id;
 	}
