+++ src/main/java/ch/hearc/zookeeper/entity/UserRole.java
@@ -14,7 +14,7 @@ public class UserRole
 	@Id
 	@GeneratedValue(strategy=GenerationType.AUTO)
 	@Column
-	private Integer id;
+	private long id;
 	
 	@Column
 	private String name;
@@ -24,7 +24,7 @@ public class UserRole
 		
 	}
 	
-	public Integer getId() 
+	public long getId() 
 	{
 		return id;
 	}
@@ -39,7 +39,7 @@ public class UserRole
 		this.name = name;
 	}
 	
-	public void setId(Integer id) 
+	public void setId(long id) 
 	{
 		this.id = id;
 	}
